add auth middleware but we do not use it yet to protect routes until we have the token managment logic also in the front end
