+++ auth/middleware.js
@@ -1,0 +1,29 @@
+const User = require("../models").user;
+const { toData } = require("../auth/jwt");
+
+async function authMiddleware(req, res, next) {
+  const auth =
+    req.headers.authorization && req.headers.authorization.split(" ");
+  if (auth && auth[0] === "Bearer" && auth[1]) {
+    try {
+      const data = toData(auth[1]);
+      const user = await User.findByPk(data.userId);
+      if (!user) {
+        res.status(404).send("No user found");
+      } else {
+        req.user = user;
+        next();
+      }
+    } catch (error) {
+      res.status(400).send({
+        message: `Error ${error.name}: ${error.message}`,
+      });
+    }
+  } else {
+    res.status(401).send({
+      message: "Please supply some valid credentials",
+    });
+  }
+}
+
+module.exports = authMiddleware;
