+++ server.js
@@ -1,13 +1,17 @@
 const { sequelize } = require("./models");
 const express = require("express");
-const productRouter = require("./productRouter");
+const productRouter = require("./routers/productRouter");
 const AuthRouter = require("./routers/auth");
-const userRouter = require("./userRouter");
+const userRouter = require("./routers/userRouter");
 const cors = require("cors");
+const authMiddleware = require("./auth/middleware");
 const app = express();
 const PORT = 4000;
 app.use(express.json());
 app.use(cors());
+
+// authentication is working properly but we cant use it until we have proper token management logic in the FE
+// app.use("/products", authMiddleware, productRouter);
 
 app.use("/products", productRouter);
 app.use("/auth", AuthRouter);
